--- vlmeval/vlm/vita/__init__.py
@@ -1,5 +1,0 @@
-from .vita import VITA
-from .vita_mistral import VITAMistral
-from .vita_qwen2 import VITAQwen2
-
-__all__ = ['VITA']
